feat(hooks): added support for hooks to perform additional setup
For many of the Launcher Boosters we need additional setup besides that what
is provided by the Jenkinsfiles in this repository. They are very specific
to each booster so a new system is introduced whereby a booster can specify
extra work that needs to be done for each environment where the booster will
be run.

To enable this a booster must add a `.openshift/Jenkinsfile.setup` file
containing two methods (both of these must exist, even if only one is
actually used!): "setupEnvironmentPre()" and "setupEnvironmentPost()".
Both have a string parameter which takes the name of the environment where
the booster will be installed. The "Pre" function will be called _before_
the actual creation of the booster and the "Post" one will be called
_afterwards_. The file also requires a single `return this` line.